pyproject.toml
[tool.poetry]
name = "kafka-practice"
version = "0.1.0"
description = ""
authors = ["3RB16 <muhammad3rbi@gmail.com>"]
readme = "README.md"
packages = [{include = "kafka_practice"}]

[tool.poetry.dependencies]
python = "^3.11"
django = "^4.2"
djangorestframework = "^3.14.0"
flake8-docstrings = "^1.7.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
